requirements.txt
discord.py
requests
bs4
psycopg2
praw